django-truck-load/truck_load/requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
alembic==1.0.6
asn1crypto==0.24.0
autoenv==1.0.0
Babel==2.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
csvkit==1.0.3
dbfread==2.0.7
Django==1.11
djangorestframework==3.6.2
et-xmlfile==1.0.1
Flask==1.0.2
Flask-API==1.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.7
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
leather==0.3.3
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==1.3.13
numpy==1.15.4
openpyxl==2.5.10
pandas==0.23.4
parsedatetime==2.4
pipenv==2018.5.18
psycopg2==2.7.6.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.5
python-editor==1.0.3
python-slugify==1.2.6
pytimeparse==1.1.8
pytz==2018.7
requests==2.19.1
six==1.12.0
SQLAlchemy==1.2.16
Unidecode==1.0.23
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
xlrd==1.1.0
